+++ src/widget/findonwebmenuservices/findonlastfmmenu.cpp
