+++ src/widget/findonwebmenuservices/findonsoundcloudmenu.cpp
